+++ TicketComment/TicketcommentRepository.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace HelpDesk.TicketComment;
 
-public class TicketCommentRepository
+public class TicketCommentRepository : ITicketCommentRepository
 {
     private readonly ApplicationDbContext _context;
 
@@ -20,22 +20,29 @@ public class TicketCommentRepository
             .ToListAsync();
     }
 
-    public async Task<TicketCommentModel?> GetByIdAsync(int id)
+    public async Task<TicketCommentModel> GetByIdAsync(int id)
     {
         return await _context.TicketComments
             .Include(tc => tc.User)
-            .FirstOrDefaultAsync(tc => tc.Id == id);
+            .FirstOrDefaultAsync(tc => tc.Id == id) ?? throw new KeyNotFoundException("Comment not found");
     }
 
-    public async Task AddAsync(TicketCommentModel comment)
+    public async Task<TicketCommentModel> CreateAsync(TicketCommentModel comment)
     {
         await _context.TicketComments.AddAsync(comment);
         await _context.SaveChangesAsync();
+        return comment;
     }
 
-    public async Task DeleteAsync(TicketCommentModel comment)
+    public async Task<bool> DeleteAsync(int id)
     {
+        var comment = await _context.TicketComments.FindAsync(id);
+        if (comment == null)
+        {
+            return false;
+        }
         _context.TicketComments.Remove(comment);
         await _context.SaveChangesAsync();
+        return true;
     }
 }
